refactor(taiko-client-rs): split oversized modules and simplify network/rest flows
