+++ packages/taiko-client-rs/Cargo.toml
@@ -77,13 +77,13 @@ jsonrpsee = { version = "0.26", features = ["server"] }
 k256 = { version = "0.13", default-features = false, features = ["arithmetic"] }
 
 # taiko
-alethia-reth-consensus = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-consensus", rev = "41fac9fd014a9944439aa2c2fbae56f9f1a627f3" }
-alethia-reth-evm = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-evm", rev = "41fac9fd014a9944439aa2c2fbae56f9f1a627f3" }
-alethia-reth-primitives = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-primitives", rev = "41fac9fd014a9944439aa2c2fbae56f9f1a627f3", features = [
+alethia-reth-consensus = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-consensus", rev = "a575e625374de3403df8517069626d007102cfc8" }
+alethia-reth-evm = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-evm", rev = "a575e625374de3403df8517069626d007102cfc8" }
+alethia-reth-primitives = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-primitives", rev = "a575e625374de3403df8517069626d007102cfc8", features = [
   "serde",
 ] }
-alethia-reth-rpc = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-rpc", rev = "41fac9fd014a9944439aa2c2fbae56f9f1a627f3" }
-reth-ipc = { git = "https://github.com/paradigmxyz/reth", tag = "v1.10.0" }
+alethia-reth-rpc = { git = "https://github.com/taikoxyz/alethia-reth", package = "alethia-reth-rpc", rev = "a575e625374de3403df8517069626d007102cfc8" }
+reth-ipc = { git = "https://github.com/paradigmxyz/reth", tag = "v1.11.1" }
 
 # types
 anyhow = "1"
