+++ packages/taiko-client-rs/crates/driver/src/derivation/pipeline/shasta/pipeline/state.rs
@@ -16,9 +16,11 @@ pub(super) struct ParentState {
     /// Anchor block number advertised by the parent block.
     pub(super) anchor_block_number: u64,
     /// Time delta between the parent and grandparent blocks.
-    pub(super) parent_block_time: u64,
+    pub(super) parent_block_time_delta_secs: u64,
     /// Timestamp when the Shasta fork is expected to activate.
     pub(super) shasta_fork_timestamp: u64,
+    /// Chain-specific minimum base fee used by EIP-4396 clamping.
+    pub(super) min_base_fee_to_clamp: u64,
 }
 
 impl ParentState {
@@ -31,18 +33,20 @@ impl ParentState {
         header: Header,
         anchor_block_number: u64,
     ) -> Result<Self, DerivationError> {
-        if header.number != self.next_block_number() {
+        let expected_block_number = self.next_block_number();
+        if header.number != expected_block_number {
             return Err(DerivationError::UnexpectedBlockNumber {
-                expected: self.next_block_number(),
+                expected: expected_block_number,
                 actual: header.number,
             });
         }
 
         Ok(Self {
-            parent_block_time: header.timestamp.saturating_sub(self.header.timestamp),
+            parent_block_time_delta_secs: header.timestamp.saturating_sub(self.header.timestamp),
             header,
             anchor_block_number,
             shasta_fork_timestamp: self.shasta_fork_timestamp,
+            min_base_fee_to_clamp: self.min_base_fee_to_clamp,
         })
     }
 
@@ -61,11 +65,20 @@ impl ParentState {
             });
         }
 
-        Ok(if self.header.number == 0 {
-            SHASTA_INITIAL_BASE_FEE
-        } else {
-            calculate_next_block_eip4396_base_fee(&self.header, self.parent_block_time)
-        })
+        if self.header.number == 0 {
+            return Ok(SHASTA_INITIAL_BASE_FEE);
+        }
+
+        let parent_base_fee_per_gas =
+            self.header.base_fee_per_gas.ok_or(DerivationError::MissingParentBaseFee {
+                parent_block_number: self.header.number,
+            })?;
+        Ok(calculate_next_block_eip4396_base_fee(
+            &self.header,
+            self.parent_block_time_delta_secs,
+            parent_base_fee_per_gas,
+            self.min_base_fee_to_clamp,
+        ))
     }
 
     /// Build the validation context used to sanity-check manifest contents.
