+++ packages/taiko-client-rs/crates/proposer/src/proposer.rs
@@ -1,7 +1,7 @@
 //! Core proposer implementation for submitting block proposals.
 
 use alethia_reth_consensus::eip4396::{
-    SHASTA_INITIAL_BASE_FEE, calculate_next_block_eip4396_base_fee,
+    MIN_BASE_FEE, SHASTA_INITIAL_BASE_FEE, calculate_next_block_eip4396_base_fee,
 };
 use alethia_reth_primitives::{
     decode_shasta_proposal_id,
@@ -45,12 +45,12 @@ use crate::{
 use alloy_provider::RootProvider;
 
 /// Type alias for batches of transaction lists fetched from the txpool.
-pub type TransactionsLists = Vec<Vec<Transaction>>;
+pub type TransactionLists = Vec<Vec<Transaction>>;
 
 /// Parameters captured from engine mode payload building.
 /// These ensure consistency between the anchor transaction and the block manifest.
 #[derive(Debug, Clone, Copy)]
-pub struct EnginePayloadParams {
+pub struct EngineBuildContext {
     /// The L1 block number used for the anchor transaction.
     pub anchor_block_number: u64,
     /// The timestamp used for the payload.
@@ -196,7 +196,7 @@ impl Proposer {
     }
 
     /// Fetch transaction pool content from the L2 execution engine.
-    async fn fetch_pool_content(&self) -> Result<TransactionsLists> {
+    async fn fetch_pool_content(&self) -> Result<TransactionLists> {
         let base_fee_u64 = u64::try_from(self.calculate_next_shasta_block_base_fee().await?)
             .map_err(|_| ProposerError::BaseFeeOverflow)?;
 
@@ -248,18 +248,26 @@ impl Proposer {
         }
 
         // Calculate the parent block time by subtracting its timestamp from its parent's timestamp.
-        let parent_block_time = parent.header.timestamp -
-            self.rpc_provider
-                .l2_provider
-                .get_block_by_number(BlockNumberOrTag::Number(parent.number() - 1))
-                .await?
-                .ok_or_else(|| ProposerError::ParentBlockNotFound(parent.number() - 1))?
-                .header
-                .timestamp;
+        let parent_number = parent.number();
+        let grandparent_number = parent_number.saturating_sub(1);
+        let grandparent = self
+            .rpc_provider
+            .l2_provider
+            .get_block_by_number(BlockNumberOrTag::Number(grandparent_number))
+            .await?
+            .ok_or(ProposerError::ParentBlockNotFound(grandparent_number))?;
+        let parent_block_time_delta_secs =
+            parent.header.timestamp.saturating_sub(grandparent.header.timestamp);
+        let parent_base_fee_per_gas =
+            parent.header.inner.base_fee_per_gas.ok_or(ProposerError::MissingParentBaseFee {
+                parent_block_number: parent_number,
+            })?;
 
         Ok(U256::from(calculate_next_block_eip4396_base_fee(
             &parent.header.inner,
-            parent_block_time,
+            parent_block_time_delta_secs,
+            parent_base_fee_per_gas,
+            MIN_BASE_FEE,
         )))
     }
 
@@ -305,7 +313,7 @@ impl Proposer {
     async fn build_payload_attributes(
         &self,
         parent: &Block,
-    ) -> Result<(TaikoPayloadAttributes, EnginePayloadParams)> {
+    ) -> Result<(TaikoPayloadAttributes, EngineBuildContext)> {
         let block_number = parent.number() + 1;
         let timestamp = current_unix_timestamp();
 
@@ -379,20 +387,21 @@ impl Proposer {
             anchor_transaction: Some(Bytes::from(anchor_tx.encoded_2718())),
         };
 
-        let engine_params = EnginePayloadParams {
-            anchor_block_number,
-            timestamp,
-            gas_limit: parent.header.gas_limit,
-        };
-
-        Ok((payload_attributes, engine_params))
+        Ok((
+            payload_attributes,
+            EngineBuildContext {
+                anchor_block_number,
+                timestamp,
+                gas_limit: parent.header.gas_limit,
+            },
+        ))
     }
 
     /// Fetch transactions using Engine API (FCU + get_payload).
     /// In engine mode, we use forkchoice_updated to trigger payload building
     /// with tx_list: None, then retrieve the built payload to extract transactions.
     /// Returns the transactions and the engine payload parameters used.
-    async fn fetch_payload_transactions(&self) -> Result<(TransactionsLists, EnginePayloadParams)> {
+    async fn fetch_payload_transactions(&self) -> Result<(TransactionLists, EngineBuildContext)> {
         // Build forkchoice state and get the head block to use as parent.
         let (forkchoice_state, parent) = self.build_forkchoice_state().await?;
