+++ packages/taiko-client-rs/crates/test-harness/src/transactions/base_fee.rs
@@ -1,7 +1,7 @@
 //! Base fee calculation for E2E tests.
 
 use alethia_reth_consensus::eip4396::{
-    SHASTA_INITIAL_BASE_FEE, calculate_next_block_eip4396_base_fee,
+    MIN_BASE_FEE, SHASTA_INITIAL_BASE_FEE, calculate_next_block_eip4396_base_fee,
 };
 use alloy_eips::BlockNumberOrTag;
 use alloy_provider::Provider;
@@ -32,14 +32,10 @@ pub async fn compute_next_block_base_fee<P>(provider: &P, parent_block_number: u
 where
     P: Provider + Send + Sync,
 {
-    let get_block = |n| async move {
-        provider
-            .get_block_by_number(BlockNumberOrTag::Number(n))
-            .await?
-            .ok_or_else(|| anyhow!("missing block {n}"))
-    };
-
-    let parent_block = get_block(parent_block_number).await?;
+    let parent_block = provider
+        .get_block_by_number(BlockNumberOrTag::Number(parent_block_number))
+        .await?
+        .ok_or_else(|| anyhow!("missing block {parent_block_number}"))?;
     let parent_header = parent_block.header.inner;
 
     // Genesis block case: return initial base fee.
@@ -47,7 +43,19 @@ where
         return Ok(SHASTA_INITIAL_BASE_FEE);
     }
 
-    let grandparent = get_block(parent_block_number.saturating_sub(1)).await?;
+    let grandparent_block_number = parent_block_number.saturating_sub(1);
+    let grandparent = provider
+        .get_block_by_number(BlockNumberOrTag::Number(grandparent_block_number))
+        .await?
+        .ok_or_else(|| anyhow!("missing block {grandparent_block_number}"))?;
     let time_delta = parent_header.timestamp.saturating_sub(grandparent.header.inner.timestamp);
-    Ok(calculate_next_block_eip4396_base_fee(&parent_header, time_delta))
+    let parent_base_fee_per_gas = parent_header
+        .base_fee_per_gas
+        .ok_or_else(|| anyhow!("parent block {parent_block_number} missing base fee"))?;
+    Ok(calculate_next_block_eip4396_base_fee(
+        &parent_header,
+        time_delta,
+        parent_base_fee_per_gas,
+        MIN_BASE_FEE,
+    ))
 }
