+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/network/mod.rs
@@ -1,0 +1,13 @@
+//! Minimal libp2p network runtime for whitelist preconfirmation topics.
+
+mod bootnodes;
+mod event_loop;
+mod gossip;
+pub(super) mod inbound;
+mod runtime;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+pub(crate) use self::types::{NetworkCommand, NetworkEvent, WhitelistNetwork};
