+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/rest_handler/compression.rs
@@ -1,0 +1,44 @@
+//! Compression and provider-error mapping helpers for REST handling.
+
+use super::*;
+
+/// Decompress a zlib-compressed transaction list.
+pub(super) fn decompress_tx_list(bytes: &[u8]) -> Result<Vec<u8>> {
+    if bytes.len() > MAX_COMPRESSED_TX_LIST_BYTES {
+        return Err(WhitelistPreconfirmationDriverError::InvalidPayload(format!(
+            "compressed tx list exceeds maximum size: {} > {}",
+            bytes.len(),
+            MAX_COMPRESSED_TX_LIST_BYTES
+        )));
+    }
+
+    let decoder = flate2::read::ZlibDecoder::new(bytes);
+    let mut out = Vec::new();
+    let read_cap = MAX_DECOMPRESSED_TX_LIST_BYTES.saturating_add(1) as u64;
+    decoder.take(read_cap).read_to_end(&mut out).map_err(|err| {
+        WhitelistPreconfirmationDriverError::InvalidPayload(format!(
+            "failed to decompress tx list from payload: {err}"
+        ))
+    })?;
+
+    if out.len() > MAX_DECOMPRESSED_TX_LIST_BYTES {
+        return Err(WhitelistPreconfirmationDriverError::InvalidPayload(format!(
+            "decompressed tx list exceeds maximum size: {} > {}",
+            out.len(),
+            MAX_DECOMPRESSED_TX_LIST_BYTES
+        )));
+    }
+
+    if out.is_empty() {
+        return Err(WhitelistPreconfirmationDriverError::InvalidPayload(
+            "decompressed tx list is empty".to_string(),
+        ));
+    }
+
+    Ok(out)
+}
+
+/// Convert a provider error into a driver error.
+pub(super) fn provider_err(err: impl std::fmt::Display) -> WhitelistPreconfirmationDriverError {
+    WhitelistPreconfirmationDriverError::Rpc(rpc::RpcClientError::Provider(err.to_string()))
+}
